+++ src/device/tsp.script/Script/SessionBaseExtensions/FirmwareManager-export.cs
@@ -32,7 +32,7 @@ public static partial class FirmwareManager
                 throw new cc.isr.VI.Pith.NativeException( $"Failed fetching script source for {script.FirmwareScript.Name}" );
 
             bool isByteCode = session.IsByteCodeScript( scriptSource );
-            ScriptFileFormats fileFormat = FirmwareScriptBase.BuildScriptFileFormat( isByteCode, compress );
+            ScriptFormats fileFormat = FirmwareScriptBase.BuildScriptFileFormat( isByteCode, compress );
             string fileTitle = cc.isr.VI.Tsp.Script.ScriptInfo.BuildScriptFileTitle( script.FirmwareScript.Name, fileFormat,
                  script.FirmwareScript.FirmwareVersion, script.FirmwareScript.ModelMask, script.FirmwareScript.ModelMajorVersion );
             if ( !script.Node.IsController )